requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
sentence-transformers
faiss-cpu
openai